refactorin errors

+++ src/main/java/app/disney/domain/model/AppUser.java
@@ -24,7 +24,7 @@ public class AppUser implements UserDetails {
     @Id()
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id", updatable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "username", nullable = false)
     private String userName;
@@ -60,11 +60,7 @@ public class AppUser implements UserDetails {
         return Objects.hash(id);
     }
 
-    /**
-     * Returns the authorities granted to the user. Cannot return <code>null</code>.
-     *
-     * @return the authorities, sorted by natural key (never <code>null</code>)
-     */
+
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         if (this.role != null) {
@@ -73,58 +69,31 @@ public class AppUser implements UserDetails {
         return Collections.emptySet();
     }
 
-    /**
-     * Returns the username used to authenticate the user. Cannot return
-     * <code>null</code>.
-     *
-     * @return the username (never <code>null</code>)
-     */
+
     @Override
     public String getUsername() {
         return this.email;
     }
 
-    /**
-     * Indicates whether the user's account has expired. An expired account cannot be
-     * authenticated.
-     *
-     * @return <code>true</code> if the user's account is valid (ie non-expired),
-     * <code>false</code> if no longer valid (ie expired)
-     */
+
     @Override
     public boolean isAccountNonExpired() {
         return true;
     }
 
-    /**
-     * Indicates whether the user is locked or unlocked. A locked user cannot be
-     * authenticated.
-     *
-     * @return <code>true</code> if the user is not locked, <code>false</code> otherwise
-     */
+
     @Override
     public boolean isAccountNonLocked() {
         return true;
     }
 
-    /**
-     * Indicates whether the user's credentials (password) has expired. Expired
-     * credentials prevent authentication.
-     *
-     * @return <code>true</code> if the user's credentials are valid (ie non-expired),
-     * <code>false</code> if no longer valid (ie expired)
-     */
+
     @Override
     public boolean isCredentialsNonExpired() {
         return true;
     }
 
-    /**
-     * Indicates whether the user is enabled or disabled. A disabled user cannot be
-     * authenticated.
-     *
-     * @return <code>true</code> if the user is enabled, <code>false</code> otherwise
-     */
+
     @Override
     public boolean isEnabled() {
         return true;
