+++ src/main/java/app/disney/domain/repository/IRoleRepository.java
@@ -1,14 +1,9 @@
 package app.disney.domain.repository;
 
+import app.disney.domain.model.AppRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import app.disney.domain.model.AppRole;
-
 @Repository
-public interface IRoleRepository extends JpaRepository<AppRole, Integer> {
-
-	
-
-	
+public interface IRoleRepository extends JpaRepository<AppRole, Long> {
 }
